% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anaconda.R
\name{check_anaconda}
\alias{check_anaconda}
\title{Checks if Anaconda is installed or not}
\usage{
check_anaconda()
}
\value{
Exit code of checking conda version
}
\description{
Checks if Anaconda is installed or not
}
