% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anaconda.R
\name{download_anaconda}
\alias{download_anaconda}
\title{Download and install Anaconda in Google Colab}
\usage{
download_anaconda()
}
\description{
Downloads and installs Anaconda. Stops runtime to restart it.
This function only needs to be run once for every Colab
runtime session.
}
