% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_drive}
\alias{download_drive}
\title{Downloads a file or files in folder in Google Drive into Google Colab}
\usage{
download_drive(drive_link)
}
\arguments{
\item{drive_link}{A string of the Drive file or folder link.
Make sure that the file or folder is public.}
}
\description{
Downloads one or more files from a Google Drive repository into Google Colab,
assuming that the files or folder they are stored in are publicly
available. Note: when downloading files from a Drive folder,
ensure that there are no sub-folders stored in the folder
itself; there can only be files inside.
}
