% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{generate_alerts}
\alias{generate_alerts}
\title{Generate Google Colab alerts}
\usage{
generate_alerts(width = 1000, height = 500, embedded = TRUE)
}
\arguments{
\item{width}{A numeric value of the width to display the embedded webpage.}

\item{height}{A numeric value of the height to display the embedded webpage.}

\item{embedded}{A boolean value on whether or not to display an embedded version
of the alert webpage. If set to false, a link to the webpage is printed instead.}
}
\description{
Display a link or an embedded webpage that helps generate Colab alerts.
For more details, view this tutorial notebook: coming soon!
}
