% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\name{generate_lda_plot_link}
\alias{generate_lda_plot_link}
\title{Plot LDA models from text2vec in Google Colab}
\usage{
generate_lda_plot_link()
}
\description{
Generates a link to use for text2vec LDA plots in Google Colab
}
