% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\name{lda_plot_text2vec}
\alias{lda_plot_text2vec}
\title{Plot LDA models from text2vec in Google Colab}
\usage{
lda_plot_text2vec(lda_model)
}
\arguments{
\item{lda_model}{A lda model object from text2vec.}
}
\description{
Plot LDA models from text2vec in Google Colab
}
