% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text2vec.R
\name{plot_lda_text2vec}
\alias{plot_lda_text2vec}
\title{Plot LDA models from text2vec in Google Colab}
\usage{
plot_lda_text2vec(lda_model, embedded = TRUE, port = 4321)
}
\arguments{
\item{lda_model}{A lda model object from text2vec.}

\item{embedded}{A boolean value on whether or not an embedded
HTML page is displayed. Otherwise, a link is generated.}

\item{port}{A port number to host tunneled server.
Default value is 4321.}
}
\description{
Allows LDA models from the text2vec package to be plotted
in Google Colab. A domain is generated to view either as
embedded or as a standalone link.
}
