% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anaconda.R
\name{setup_anaconda}
\alias{setup_anaconda}
\title{Setup Anaconda to work in Google Colab}
\usage{
setup_anaconda(env = "r-reticulate")
}
\arguments{
\item{env}{A string name for the conda environment.}
}
\description{
Downloads and installs Anaconda in Google Colab. Also
uses a user specified environment variable.
}
