% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNLP.R
\name{setup_cleanNLP}
\alias{setup_cleanNLP}
\title{Setup cleanNLP to work in Google Colab}
\usage{
setup_cleanNLP(env = "r-reticulate")
}
\arguments{
\item{env}{A string name for the conda environment.
Default value is 'r-reticulate'.}
}
\description{
Setups and initalizes the cleanNLP package to work in Google Colab.
Please ensure you have installed Anaconda using setup_anaconda
first and have installed the spacyr package.
}
