% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacyr.R
\name{setup_spacyr}
\alias{setup_spacyr}
\title{Setup spacyr to work in Google Colab}
\usage{
setup_spacyr(env = "r-reticulate", lang_models = "en_core_web_sm")
}
\arguments{
\item{env}{A string name for the conda environment.
Default is called 'r-reticulate'.}

\item{lang_models}{The language package for loading spaCy.
Default is 'en_core_web_sm' (English). For more models
and languages, go to spacy.io/usage/models.}
}
\description{
Setups and initalizes the spacyr package to work in Google Colab.
Please ensure you have installed Anaconda using setup_anaconda
first and have installed the spacyr package.
}
