% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alerts.R
\name{generate_alerts}
\alias{generate_alerts}
\title{Generate Google Colab alerts}
\usage{
generate_alerts(embedded = TRUE, width = 800, height = 500)
}
\arguments{
\item{width}{A numeric value of the width to display the embedded webpage.}

\item{height}{A numeric value of the height to display the embedded webpage.}

\item{embeded}{A boolean value on whether or not to display an embedded version
of the alert webpage. If set to false, a link to the webpage is printed instead.}
}
\description{
Display a link or an embedded webpage that helps generate Colab alerts.
}
