% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anaconda.R
\name{setup_anaconda}
\alias{setup_anaconda}
\title{Setup an Anaconda environment}
\usage{
setup_anaconda(env = "r-reticulate")
}
\arguments{
\item{env}{A string name for the conda environment.
Default value is 'r-reticulate'.}
}
\description{
Setup an Anaconda environment to work in Google Colab.
}
